+++ CHANGELOG.rst
@@ -8,6 +8,16 @@ inverse chronological order.
 Changes since version 5.0
 -------------------------
 
+- Sparse module
+
+  The new :mod:`nutils.sparse` module introduces a data type and a suite of
+  manipulation methods for arbitrary dimensional sparse data. The existing
+  integrate and integral methods now create data of this type under the hood,
+  and then :func:`nutils.sparse.convert` it to a scalar, Numpy array or
+  :class:`nutils.matrix.Matrix` upon return. In order to receive the sparse
+  objects instead use the new :func:`nutils.sample.Sample.integrate_sparse` or
+  :func:`nutils.sample.eval_integrals_sparse`.
+
 - External dependency for parsing gmsh files
 
   The :func:`nutils.mesh.gmsh` method now depends on the external `meshio